+++ packages/launchpad/src/encoding.spec.ts
@@ -1,4 +1,8 @@
-import { sortedJsonStringify } from "./encoding";
+/* eslint-disable @typescript-eslint/naming-convention */
+import { coin, coins } from "./coins";
+import { makeSignDoc, sortedJsonStringify } from "./encoding";
+import { MsgDelegate, MsgSend } from "./msgs";
+import { faucet, launchpad, makeRandomAddress } from "./testutils.spec";
 
 describe("encoding", () => {
   describe("sortedJsonStringify", () => {
@@ -45,6 +49,81 @@ describe("encoding", () => {
       expect(sortedJsonStringify([1, 2, { b: { z: true, x: true, y: true }, a: true, c: true }, 4])).toEqual(
         `[1,2,{"a":true,"b":{"x":true,"y":true,"z":true},"c":true},4]`,
       );
+    });
+  });
+
+  describe("makeSignDoc", () => {
+    it("works", () => {
+      const chainId = "testspace-12";
+      const msg1: MsgDelegate = {
+        type: "cosmos-sdk/MsgDelegate",
+        value: {
+          delegator_address: faucet.address,
+          validator_address: launchpad.validator.address,
+          amount: coin(1234, "ustake"),
+        },
+      };
+      const msg2: MsgSend = {
+        type: "cosmos-sdk/MsgSend",
+        value: {
+          from_address: faucet.address,
+          to_address: makeRandomAddress(),
+          amount: coins(1234567, "ucosm"),
+        },
+      };
+      const fee = {
+        amount: coins(2000, "ucosm"),
+        gas: "180000", // 180k
+      };
+      const memo = "Use your power wisely";
+      const accountNumber = 15;
+      const sequence = 16;
+
+      const signDoc = makeSignDoc([msg1, msg2], fee, chainId, memo, accountNumber, sequence);
+      expect(signDoc).toEqual({
+        msgs: [msg1, msg2],
+        fee: fee,
+        chain_id: chainId,
+        account_number: accountNumber.toString(),
+        sequence: sequence.toString(),
+        memo: memo,
+      });
+    });
+
+    it("works with undefined memo", () => {
+      const chainId = "testspace-12";
+      const msg1: MsgDelegate = {
+        type: "cosmos-sdk/MsgDelegate",
+        value: {
+          delegator_address: faucet.address,
+          validator_address: launchpad.validator.address,
+          amount: coin(1234, "ustake"),
+        },
+      };
+      const msg2: MsgSend = {
+        type: "cosmos-sdk/MsgSend",
+        value: {
+          from_address: faucet.address,
+          to_address: makeRandomAddress(),
+          amount: coins(1234567, "ucosm"),
+        },
+      };
+      const fee = {
+        amount: coins(2000, "ucosm"),
+        gas: "180000", // 180k
+      };
+      const accountNumber = 15;
+      const sequence = 16;
+
+      const signDoc = makeSignDoc([msg1, msg2], fee, chainId, undefined, accountNumber, sequence);
+      expect(signDoc).toEqual({
+        msgs: [msg1, msg2],
+        fee: fee,
+        chain_id: chainId,
+        account_number: accountNumber.toString(),
+        sequence: sequence.toString(),
+        memo: "",
+      });
     });
   });
 });